feat: redis 사용 설정
